+++ lib/terraspace_bundler/cli/runner.rb
@@ -1,7 +1,0 @@
-class TerraspaceBundler::CLI
-  class Runner < Base
-    def run
-      TB::Installer.new(@options).run
-    end
-  end
-end