+++ lib/terraspace_bundler/dsl/concern.rb
@@ -1,7 +1,0 @@
-class TerraspaceBundler::Dsl
-  module Concern
-    def dsl
-      TB.dsl
-    end
-  end
-end