+++ lib/terraspace_bundler/exporter/stacks/base.rb
@@ -1,0 +1,5 @@
+class TerraspaceBundler::Exporter::Stacks
+  class Base < TerraspaceBundler::Exporter::Base
+    include TB::Util::Logging
+  end
+end