+++ spec/fixtures/rewrite/example-subfolder.tf
@@ -1,0 +1,3 @@
+module "compute_instance" {
+  source            = "../../../modules/compute_instance"
+}