+++ spec/fixtures/rewrite/module-subfolder.tf
@@ -1,0 +1,3 @@
+module "iam-user" {
+  source = "../../modules/iam-user"
+}