+++ spec/fixtures/rewrite/no-leading.tf
@@ -1,0 +1,3 @@
+module "consul" {
+source="../.." # no leading space
+}