+++ spec/fixtures/rewrite/spaces.tf
@@ -1,0 +1,3 @@
+module "consul" {
+  source     =   "../.." # lots of extra spaces
+}